feat(wip): refactor task and api into crate::ipfs
