+++ hermes/bin/src/runtime_extensions/hermes/ipfs/host.rs
@@ -1,6 +1,6 @@
 //! IPFS host implementation for WASM runtime.
 
-use super::state::{
+use crate::ipfs::{
     hermes_ipfs_add_file, hermes_ipfs_content_validate, hermes_ipfs_evict_peer,
     hermes_ipfs_get_dht_value, hermes_ipfs_get_file, hermes_ipfs_pin_file, hermes_ipfs_publish,
     hermes_ipfs_put_dht_value, hermes_ipfs_subscribe, hermes_ipfs_unpin_file,
