+++ hermes/bin/src/runtime_extensions/hermes/ipfs/mod.rs
@@ -1,7 +1,6 @@
 //! Hermes IPFS runtime extension.
-mod event;
+pub(crate) mod event;
 mod host;
-mod state;
 
 /// Advise Runtime Extensions of a new context
 pub(crate) fn new_context(_ctx: &crate::runtime_context::HermesRuntimeContext) {}
